Implemented tupled and recursive functions.
Added support for tuples and implemented multi-arity
functions as tupled functions. Added optimization that
does automatic boxing and unboxing of tuples so in the
generated code we can still have multi-arity functions.

Added support for recursive functions using one of
techniques outlined in tests augmented with support for
side-effects.

Tupled functions are a Base trait, while Recursive functions
are an implementation Exp trait. They can be used separately
or mixed-in with one another.