Update test-cases after update to 2.10.0-M2.
Those test-cases are broken due to the fact
that they rely on printing HashMaps that
have no contract on the order of printing
their elements. Some changes in Scala library
related to HashMaps caused this problem.

Let me repeat: those tests are broken and
should be refactored to rely on stable
sorting of pairs stored in a HashMap.