experiments with dynamic recompilation
Introduce `stable` vars. Programs are compiled assuming that stable
vars are mostly constant. Programs are specialized to the current
values but every read is replaced by a guard that will trigger
recompilation of the program if the var has changed.

The tricky part is recompiling the program on the fly, because
when we detect a var change, we are already in the middle of
executing compiled code. This is solved by storing the uncompiled
continuation as static constant. Additional care must be taken to
deal with the free vars of the continuation, which may point
to local vars of the running compiled code. We thus generate
instructions in the original code to pass these vars to
the dynamic compile call.

There are considerable degrees of freedom regarding when to
compile, what compiled code to store/throw away etc. The test
cases explore just one design.